updating precision for unit tests

+++ tests/CabbageAppTests.cpp
@@ -139,12 +139,12 @@ TEST_CASE("Test WebSocket Server functionality", "[CabbageApp]")
     int nBufferFrames = 512;
 
     // Create a dud buffer from processor to avoid segfaults in CI mode
-    float **buffer = new float *[nOutputChannels];
+    double **buffer = new double*[nOutputChannels];
     for (unsigned int ch = 0; ch < nOutputChannels; ++ch)
     {
-        buffer[ch] = new float[nBufferFrames];
+        buffer[ch] = new double[nBufferFrames];
         // Initialize to silence
-        memset(buffer[ch], 0, nBufferFrames * sizeof(float));
+        memset(buffer[ch], 0, nBufferFrames * sizeof(double));
     }
